alumno: ver lista de pruebas, ejecutar pruebas. docente evaluar pruebas

+++ database/migrations/2020_08_02_013918_create_prueba_table.php
@@ -26,7 +26,7 @@ class CreatePruebaTable extends Migration
             $table->date('fe_prueba')->nullable();
             $table->time('hh_inicio')->nullable();
             $table->time('hh_fin')->nullable();
-            $table->integer('nu_valor_total')->nullable();
+            $table->decimal('nu_valor_total', 11, 2)->nullable();
             $table->string('tx_observaciones', 100)->nullable();
             $table->integer('id_status');
             $table->integer('id_usuario');
